fix: implement correct display 10 best scores
