+++ src/components/Modal.module.css
@@ -1,53 +1,65 @@
 .modal.active {
-   width: 100%;
-   height: 100%;
-   transition: all 1.2s ease;
-   position: fixed;
-   top: 0;
-   right: 0;
-   bottom: 0;
-   left: 0;
-   background: rgba(33, 156, 60, 0.7);
-   display: flex;
-   flex-direction: column;
-   justify-content: center;
-   align-items: center;
-   color: #090e3e;
-   transform: translate(0, -200%);
-   color: #021a09;
- }
- .modal.active {
-   transform: translate(0, 0);
-   transition: all 0.6s ease;
- }
- .play_btn {
-   margin: 20px 0 30px;
-   cursor: pointer;
-   background-size: 220%;
-   border-radius: 10px;
-   box-shadow: 0 5px 5px -2px;
-   color: rgb(46, 21, 5);
-   transition: all 0.5s ease;
-   border: none;
-   outline: none;
-   padding: 10px 20px;
-   background-image: linear-gradient(45deg, #fcee21, #a9d605, #fcee21);
- }
- .play_btn span {
-   font-size: 18px;
-   text-transform: uppercase;
-   font-weight: 900;
-   color: #021a09;
- }
- .play_btn:active,
- .play_btn:hover {
-   background-position: right;
-   outline: none;
-   border: none;
-   box-shadow: 0 5px 5px -5px;
- }
- .advanced{
+  width: 100%;
+  height: 100%;
+  transition: all 1.2s ease;
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  background: rgba(33, 156, 60, 0.7);
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  color: #090e3e;
+  transform: translate(0, -200%);
+  color: #021a09;
+}
+.modal.active {
+  transform: translate(0, 0);
+  transition: all 0.6s ease;
+}
+.play_btn {
+  margin: 20px 0 30px;
+  cursor: pointer;
+  background-size: 220%;
+  border-radius: 10px;
+  box-shadow: 0 5px 5px -2px;
+  color: rgb(46, 21, 5);
+  transition: all 0.5s ease;
+  border: none;
+  outline: none;
+  padding: 10px 20px;
+  background-image: linear-gradient(45deg, #fcee21, #a9d605, #fcee21);
+}
+.play_btn span {
+  font-size: 18px;
+  text-transform: uppercase;
+  font-weight: 900;
+  color: #021a09;
+}
+.play_btn:active,
+.play_btn:hover {
+  background-position: right;
+  outline: none;
+  border: none;
+  box-shadow: 0 5px 5px -5px;
+}
+.advanced {
   width: 100px;
   display: flex;
   justify-content: space-between;
+}
+
+.attributes,
+.attributes a {
+  color: #28522e;
+}
+.attributes {
+  font-size: 10px;
+  margin-left: 10px;
+  position: fixed;
+  bottom:5px;
+  left:0;
 }
