+++ src/constants.js
@@ -21,6 +21,7 @@ const RIGHT_DIRECTION = { left: STEP, top: 0 };
 const CURRENT_DIRECTION = RIGHT_DIRECTION;
 
 const QUANTITY_BEST_RESULTS = 10;
+const ALL_SCORES  = Array.from(Array(QUANTITY_BEST_RESULTS), () => 0);
 
 export {
   FIELD_WIDTH,
@@ -39,5 +40,5 @@ export {
   DOWN_DIRECTION,
   LEFT_DIRECTION,
   RIGHT_DIRECTION,
-  QUANTITY_BEST_RESULTS
+  ALL_SCORES
 };
